refactor: rename hoofdprocess label to proces + verander to wijzig
Went over the gitbook feature passport and there was written that
```
In the design it says "hoofdproces" we are going to use "proces"
Make sure to change this everywhere
```

```
Note: in the design there is "veranderen" next to the process. But instead we will do the same "wijzig" as in the block of "algemene informatie".
```

Refs: oph-716