oph 716 archive category domain group usage check (#139)
* feat: hide the delete button for groups of a domain on admin page

Refs: oph-716

* feat: create a edit-toolbar component for process-category & domain

This is the idea for now I added this toolbar for the category

Refs: oph-716

* refactor: replace new domain button with the edit-toolbar

TODO:
- extend the toolbar with a create button
- Also use this toolbar for domains

Refs: oph-716

* feat: add labels to the action buttons but hide the text

Refs: oph-716

* feat: use the edit-toolbar for domains aswell

Domains are refreshed when one is removed

Refs: oph-716

* feat: edit-toolbar can also update the label

Refs: oph-716

* feat: add the create domain button in the category dropdown

Refs: oph-716

* refactor: create category is also done by the edit-toolbar

Updated the styling of add domain button to match the other buttons

Refs: oph-716

* feat: create label with status pill component

Can add a label or a statusUri of both

Refs: oph-716

* feat: use the label with status pill in the edit-toolbar

Refs: oph-716

* refactor: make the pill a round circle with the correct color (not sure about it)

Refs: oph-716

* feat: add process-api call to check usage

When we want to delete or archive we have to check the usage

Refs: oph-716

* feat: also check the edit group row on usage

Would like to use the toolbar in here but as the data is harder to do I'll leave it as is

Refs: oph-716

* feat: do not check usage with service but with util method through ember store

Doing it like this as we have all the information to specifically search for the used model in the frontend and not in the backend service.

Refs: oph-716

* refactor: create a conceptual-process service that will do the util method

- Removed the util as we need the store, don't like it passing it on
- Use the service in both places so they do the same logic

Refs: oph-716

* refactor: remove the status pill on the admin page

Refs: oph-716

* fix: the groups of a domain now get the og relations added when update

The domain relation of the group got removed for some reason so adding it explicitly now

Refs: oph-716

* refactor: remove unused style class

LabelWithStatusPill will not be round

Refs: oph-716

* refactor: show the usage message in an AuAlert

Because just text is not "visible enough"

Refs: oph-716

* refactor: rename hoofdprocess label to proces + verander to wijzig

Went over the gitbook feature passport and there was written that
```
In the design it says "hoofdproces" we are going to use "proces"
Make sure to change this everywhere
```

```
Note: in the design there is "veranderen" next to the process. But instead we will do the same "wijzig" as in the block of "algemene informatie".
```

Refs: oph-716

* Oph 716 | with martijns components form filter pr (#141)

* feat: create a delete modal component for inventory-process

Based on Martijns component

Refs: oph-716

* feat: create an upsert modal components

This will create and update the passed on process

Refs: oph-716